feat ResultCard Decoupled
Code was decoupled into 3 other different components - ResultCard (this one), CardResultValid, CardResultNotFound. That included decoupling of the tests.

This made Rendering and understand the code easier, as both CardResultValid and CardResultNotFound, only required to use the props passed down - with no additional alterations made.

Thus the responsibility ResultCard is to house the internal logic of which type of card will be shown to the end user.